mysite/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-tinymce==3.5.0
sqlparse==0.4.3
tzdata==2022.6
